Add copyright notice and AGPL license to GlobalObject.cs
Added a copyright notice and licensing information to the top of
the `GlobalObject.cs` file. The notice specifies that the code is
copyrighted by Kaz Nishimura in 2025 and is licensed under the
GNU Affero General Public License (AGPL) version 3 or later.
Included a disclaimer of warranties and a reference to the full
license text available at <http://www.gnu.org/licenses/>. Added
SPDX license identifier `AGPL-3.0-or-later` for clarity.