Move GlobalObject instantiation outside try block
The GlobalObject instantiation has been moved outside of the try
block to ensure it is created before any potential exceptions are
caught. This change allows the globalObject to be used in the catch
blocks. Specifically, the ScriptEngineException catch block now
writes the exception message to globalObject.Output instead of
having a TODO comment.